+++ src/ui/public/vis_maps/markers/shaded_circles.js
